Fix grace for auto-computed scales
The 'grace' for automatically computed scales was taken proportional to the data value range.
This is not appropriate if one of the min or max scales is manually set. In particular, when
you have e.g. a manual min Y scale value of 0, and every data point has the same Y value, this
caused the line to be drawed at the edge of the graph.
(see https://sourceforge.net/p/zedgraph/patches/74)